% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.diagmeta.R
\name{print.diagmeta}
\alias{print.diagmeta}
\title{Print method for diagmeta objects}
\usage{
\method{print}{diagmeta}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{diagmeta}.}

\item{digits}{Number of significant digits for printing.}

\item{\dots}{Additional arguments.}
}
\description{
Print method for objects of class \code{diagmeta}.
}
\examples{
# FENO dataset
#
data(Schneider2017)

diag1 <- diagmeta(tpos, fpos, tneg, fneg, cutpoint,
                  studlab = paste(author, year, group),
                  data = Schneider2017,
                  model = "DIDS", log.cutoff = TRUE)

diag1
}
\seealso{
\code{\link{diagmeta}} \code{\link{summary.diagmeta}}
}
\author{
Gerta Rücker \email{ruecker@imbi.uni-freiburg.de},
Susanne Steinhauser \email{susanne.steinhauser@uni-koeln.de},
Srinath Kolampally \email{kolampal@imbi.uni-freiburg.de}
}
\keyword{print}
