% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.diagmeta.R
\name{summary.diagmeta}
\alias{summary.diagmeta}
\title{Summary method for diagmeta}
\usage{
\method{summary}{diagmeta}(object, ...)
}
\arguments{
\item{object}{An object of class \code{diagmeta}.}

\item{\dots}{Additional arguments.}
}
\value{
A list with classes 'summary.diagmeta' and 'diagmeta' is
returned. The list elements are identical to a
\code{\link{diagmeta}} object.
}
\description{
Summary method for objects of class \code{diagmeta}.
}
\examples{
# FENO dataset
#
data(Schneider2017)

diag1 <- diagmeta(tpos, fpos, tneg, fneg, cutpoint,
                  studlab = paste(author, year, group),
                  data = Schneider2017,
                  log.cutoff = TRUE)

summary(diag1)

}
\seealso{
\code{\link{diagmeta}}
}
\author{
Srinath Kolampally \email{kolampal@imbi.uni-freiburg.de},
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
